refactor: refactor checkBoxGroup
