+++ src/components/inputs/CheckBoxGroup/CheckboxGroupItem.tsx
@@ -3,15 +3,15 @@
 import React from "react";
 import { CheckIcon } from "@heroicons/react/24/outline";
 
-import Label from "./Label";
+import Label from "@/components/inputs/Label";
 
-interface CheckboxGroupItemProps
+export interface CheckboxGroupItemProps
   extends React.InputHTMLAttributes<HTMLInputElement> {
   id: string;
   label: string | JSX.Element;
 }
 
-const CheckboxGroupItem = React.forwardRef<
+export const CheckboxGroupItem = React.forwardRef<
   HTMLInputElement,
   CheckboxGroupItemProps
 >(({ id, label, ...props }, ref) => (
@@ -30,35 +30,9 @@ const CheckboxGroupItem = React.forwardRef<
       <span className="flex items-center justify-center w-6 h-6 border rounded bg-base-200 border-neutral/40 transition-all [&>*]:hidden group-hover:bg-base-100 peer-checked:border-base-300 [&>*]:text-base-300 peer-checked:[&>*]:block group-hover:peer-checked:[&>*]:text-neutral-content group-hover:peer-checked:border-neutral-content">
         <CheckIcon className="hidden transition" />
       </span>
-      {label}
+      <span className="peer-checked:text-base-300">{label}</span>
     </Label>
   </div>
 ));
 
 CheckboxGroupItem.displayName = "CheckboxGroupItem";
-
-interface CheckboxGroupProps
-  extends React.InputHTMLAttributes<HTMLInputElement> {
-  options: CheckboxGroupItemProps[];
-}
-
-const CheckboxGroup = React.forwardRef<HTMLInputElement, CheckboxGroupProps>(
-  ({ options, ...props }, ref) => (
-    <div className="flex flex-col gap-y-5">
-      {options.map(({ id, value, label }) => (
-        <CheckboxGroupItem
-          key={id}
-          id={id}
-          value={value}
-          label={label}
-          {...props}
-          ref={ref}
-        />
-      ))}
-    </div>
-  ),
-);
-
-CheckboxGroup.displayName = "CheckboxGroup";
-
-export default CheckboxGroup;
